+++ gegelatilib/include/learn/learningAgent.h
@@ -137,9 +137,14 @@ namespace Learn {
               tpg(factory.createTPGGraph(env)), params{p},
               archive(p.archiveSize, p.archivingProbability)
         {
-            // override the number of actions from the parameters.
-            this->params.mutation.tpg.nbActions =
-                this->learningEnvironment.getNbActions();
+
+            // override the number of initial roots if set to 0
+            if (this->params.mutation.tpg.initNbRoots == 0) {
+                this->params.mutation.tpg.initNbRoots = std::max(
+                    (size_t)floor((1 - this->params.ratioDeletedRoots) *
+                                  (double)params.mutation.tpg.nbRoots),
+                    (size_t)this->learningEnvironment.getNbActions());
+            }
         };
 
         /// Default destructor for polymorphism