+++ gegelatilib/include/mutator/mutationParameters.h
@@ -46,14 +46,6 @@ namespace Mutator {
      */
     typedef struct TPGParameters
     {
-        /// JSon comment
-        inline static const std::string nbActionsComment =
-            "// Number of TPGAction vertex of the initialized TPGGraph.\n"
-            "// This parameter is generally automatically set by the "
-            "LearningEnvironment.\n"
-            "// /* \"nbActions\" : 0,*/ // Commented by default";
-        /// Number of TPGAction vertex of the initialized TPGGraph.
-        size_t nbActions = 0;
 
         /// JSon comment
         inline static const std::string nbRootsComment =
@@ -62,6 +54,16 @@ namespace Mutator {
             "// \"nbRoots\" : 100, // Default value";
         /// Number of root TPGTeams to maintain when populating the TPGGraph
         size_t nbRoots = 100;
+
+        /// JSon comment
+        inline static const std::string initNbRootsComment =
+            "// Number of root TPGTeams at the initialisation of a "
+            "TPGGraph.\n"
+            "// If 0, if will be init to the number of surviving roots\n"
+            "// \"nbRoots\" : 0, // Default value";
+        /// Number of root TPGTeams at the initialisation of a TPGGraph
+        /// If 0, if will be init to the number of surviving roots
+        size_t initNbRoots = 0;
 
         /// JSon comment
         inline static const std::string maxInitOutgoingEdgesComment =
@@ -176,6 +178,13 @@ namespace Mutator {
             "// \"pConstantMutation\" : 0.5, // Default value";
         /// Probability of each constant to be mutated
         double pConstantMutation = 0.5;
+
+        /// JSon comment
+        inline static const std::string pNewProgramComment =
+            "// Probability of creating a new program.\n"
+            "// \"pNewProgram\" : 0.0, // Default value";
+        /// Probability of creating a new program
+        double pNewProgram = 0.0;
 
         /// JSon comment
         inline static const std::string minConstValueComment =