+++ gegelatilib/src/file/parametersParser.cpp
@@ -101,12 +101,12 @@ void File::ParametersParser::setParameterFromString(
     Learn::LearningParameters& params, const std::string& param,
     Json::Value const& value)
 {
-    if (param == "nbActions") {
-        params.mutation.tpg.nbActions = (size_t)value.asUInt();
-        return;
-    }
     if (param == "nbRoots") {
         params.mutation.tpg.nbRoots = (size_t)value.asUInt();
+        return;
+    }
+    if (param == "initNbRoots") {
+        params.mutation.tpg.initNbRoots = (size_t)value.asUInt();
         return;
     }
     if (param == "maxInitOutgoingEdges") {
@@ -168,6 +168,10 @@ void File::ParametersParser::setParameterFromString(
     }
     if (param == "pConstantMutation") {
         params.mutation.prog.pConstantMutation = value.asDouble();
+        return;
+    }
+    if (param == "pNewProgram") {
+        params.mutation.prog.pNewProgram = value.asDouble();
         return;
     }
     if (param == "minConstValue") {
@@ -318,13 +322,13 @@ void File::ParametersParser::writeParametersToJson(
     root["mutation"]["tpg"]["maxOutgoingEdges"].setComment(
         Mutator::TPGParameters::maxOutgoingEdgesComment, Json::commentBefore);
 
-    root["mutation"]["tpg"]["nbActions"] = params.mutation.tpg.nbActions;
-    root["mutation"]["tpg"]["nbActions"].setComment(
-        Mutator::TPGParameters::nbActionsComment, Json::commentBefore);
-
     root["mutation"]["tpg"]["nbRoots"] = params.mutation.tpg.nbRoots;
     root["mutation"]["tpg"]["nbRoots"].setComment(
         Mutator::TPGParameters::nbRootsComment, Json::commentBefore);
+
+    root["mutation"]["tpg"]["initNbRoots"] = params.mutation.tpg.initNbRoots;
+    root["mutation"]["tpg"]["initNbRoots"].setComment(
+        Mutator::TPGParameters::initNbRootsComment, Json::commentBefore);
 
     root["mutation"]["tpg"]["pEdgeAddition"] =
         params.mutation.tpg.pEdgeAddition;
@@ -378,6 +382,10 @@ void File::ParametersParser::writeParametersToJson(
     root["mutation"]["prog"]["pConstantMutation"].setComment(
         Mutator::ProgramParameters::pConstantMutationComment,
         Json::commentBefore);
+
+    root["mutation"]["prog"]["pNewProgram"] = params.mutation.prog.pNewProgram;
+    root["mutation"]["prog"]["pNewProgram"].setComment(
+        Mutator::ProgramParameters::pNewProgramComment, Json::commentBefore);
 
     root["mutation"]["prog"]["pDelete"] = params.mutation.prog.pDelete;
     root["mutation"]["prog"]["pDelete"].setComment(